Add main guard, refactor, run as module
